Tweak config to remove unused modules

+++ scripts/setup/touchpad.sh
@@ -8,7 +8,6 @@ make -j4
 echo Installing libgestures
 sudo make install
 cd ..
-rm -rf libgestures
 
 echo Compiling libevdevc
 git clone https://github.com/hugegreenbug/libevdevc
@@ -17,7 +16,6 @@ make -j4
 echo Installing libevdevc
 sudo make install
 cd ..
-rm -rf libevdevc
 
 echo Compiling xf86-input-cmt
 git clone https://github.com/hugegreenbug/xf86-input-cmt
@@ -29,10 +27,17 @@ echo Installing xf86-input-cmt
 sudo make install
 
 echo Configuring
-sudo cp xorg-conf/40-touchpad-cmt.conf /usr/share/X11/xorg.conf.d/
-sudo cp xorg-conf/50-touchpad-cmt-samus.conf /usr/share/X11/xorg.conf.d/
+if [[ ! -f /usr/share/X11/xorg.conf.d/40-touchpad-cmt.conf ]]; then
+  sudo cp xorg-conf/40-touchpad-cmt.conf /usr/share/X11/xorg.conf.d/
+fi
+if [[ ! -f /usr/share/X11/xorg.conf.d/50-touchpad-cmt-samus.conf ]]; then
+  sudo cp xorg-conf/50-touchpad-cmt-samus.conf /usr/share/X11/xorg.conf.d/
+fi
 set +e
+
+
+sudo rm -rf libgestures
+sudo rm -rf libevdevc
+sudo rm -rf xf86-input-cmt
 sudo mv /usr/share/X11/xorg.conf.d/50-synaptics.conf /usr/share/X11/xorg.conf/50-synaptics.conf.old 2>/dev/null
-
-
 echo all done! - reboot or restart X
